# src/common/swagger/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "3.0.0",
        "title": "allinjob API Docs",
        "description": "올인잡 api 문서입니다"
    },
    "host": "https://allinjob.co.kr",
    "servers": [
        {
            "url": "http://localhost:4000"
        },
        {
            "url": "https://allinjob.co.kr"
        }
    ],
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "CookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "refreshToken"
            }
        }
    },
    "paths": {
        "/user/findOneUserByEmail": {
            "get": {
                "tags": ["Users"],
                "summary": "회원가입한 유저 이메일 찾아오기",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "abc@abc.com"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "응답 성공",
                        "content": {
                            "application/json": {
                                "example": {
                                    "value": {
                                        "data": {
                                            "id": "adfhttwerqwrasxczx",
                                            "email": "abc@abc.com",
                                            "provider": "kakao, naver, google",
                                            "name": "홍길동",
                                            "nickname": "닉네임",
                                            "phone": "01012345678",
                                            "profileImage": "awihqwifbjqjfbo",
                                            "major": "컴퓨터공학",
                                            "interestKeyword": [
                                                {
                                                    "id": "asdsadasdasdasdasdsef",
                                                    "userId": "adfhttwerqwrasxczx",
                                                    "interestId": "adwwafefewfsdfdsgew",
                                                    "keywordId": "asdwfwgregrerfsfsefe",
                                                    "interest": {
                                                        "id": "adwwafefewfsdfdsgew",
                                                        "interest": "공모전, 대외활동, 자격증, 어학, 인턴"
                                                    },
                                                    "keyword": {
                                                        "id": "asdwfwgregrerfsfsefe",
                                                        "keyword": "기획/아이디어, 광고/마케팅, 사진/영상/UCC, 디자인/순수예술/공예, 네이밍/슬로건, 캐릭터/만화/게임, 건축/건설/인테리어, 과학/공학, 예체능/패션, 전시/패션, 문학/시나리오, 해외, 학술, 창업, 기타, 서포터즈, 해외탐방/무료, 해외탐방/유료, 봉사단/해외, 봉사단/국내, 마케터, 기자단, 강연, 멘토링, 경영/회계/사무, 보건/의료, 문화예술/다지안/방송, 사회복지/종교, 이용/숙박/여행/오락/스포츠, 음식서비스, 기계, 정보통신, 식품/가공, 환경/에너지, TOEIC, TOEIC.speaking and writing tests, TOEFL, TOEIC bridge, TOEIC bridge.speaking and writing tests, 대기업, 중소기업, 공공기관/공기업, 외국계기업, 중견기업, 비영리단체/협회/재단, 동아리/학생자치단체"
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "provider": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "nickname": {
                                                    "type": "string"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                },
                                                "profileImage": {
                                                    "type": "string"
                                                },
                                                "major": {
                                                    "type": "string"
                                                },
                                                "interestKeyword": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string"
                                                            },
                                                            "userId": {
                                                                "type": "string"
                                                            },
                                                            "interestId": {
                                                                "type": "string"
                                                            },
                                                            "keywordId": {
                                                                "type": "string"
                                                            },
                                                            "interest": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string"
                                                                    },
                                                                    "interest": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            },
                                                            "keyword": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string"
                                                                    },
                                                                    "keyword": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "요청 형식 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "email must be an email"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "회원가입한 유저가 없는 경우",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "회원가입한 유저가 없습니다"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/findOneUserByID": {
            "get": {
                "tags": ["Users"],
                "summary": "아이디 찾기",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "홍길동"
                        },
                        "description": "이름"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "01012345678"
                        },
                        "description": "전화번호"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "응답 성공",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": {
                                        "email": "abc@abc.com",
                                        "provider": "kakao, naver, google"
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "email": {
                                                    "type": "string"
                                                },
                                                "provider": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "요청 형식 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": [
                                        "name : name must be a string",
                                        "phone : phone must be a string"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "일치하는 아이디가 없는 경우",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "일치하는 아이디가 없습니다"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/findUserProfile": {
            "get": {
                "tags": ["Users"],
                "summary": "유저 프로필 조회하기",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "응답 성공",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": {
                                        "email": "abc@abc.com",
                                        "profileImage": "profile url",
                                        "nickname": "닉네임",
                                        "interestKeyword": [
                                            {
                                                "interest": "competition / outside / intern / language / qnet",
                                                "keyword": [
                                                    "키워드1",
                                                    "키워드2"
                                                ]
                                            }
                                        ]
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "email": {
                                                    "type": "string"
                                                },
                                                "profileImage": {
                                                    "type": "string"
                                                },
                                                "nickname": {
                                                    "type": "string"
                                                },
                                                "interestKeyword": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "interest": {
                                                                "type": "string"
                                                            },
                                                            "keyword": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "string",
                                                                    "keyword": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/isNickname": {
            "get": {
                "tags": ["Users"],
                "summary": "닉네임 중복 여부 확인하기",
                "parameters": [
                    {
                        "name": "nickname",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "닉네임"
                        },
                        "description": "닉네임"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "닉네임 사용 가능",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": true
                                },
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "NoLogin": {
                                        "summary": "닉네임 중복되는 경우",
                                        "value": {
                                            "error": "이미 사용중인 닉네임입니다"
                                        }
                                    },
                                    "TypeError": {
                                        "summary": "닉네임 형식이 잘못된 경우",
                                        "value": {
                                            "error": "nickname: nickname must match /^[a-zA-Z가-힣]+$/ regular expression"
                                        }
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "string"
                                        },
                                        {
                                            "type": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/getCount": {
            "get": {
                "tags": ["Users"],
                "summary": "유저의 각 활동내역별 점수 계산",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "응답 성공",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": [
                                        {
                                            "competition": {
                                                "barType": "competition",
                                                "percent": 0
                                            },
                                            "outside": {
                                                "barType": "outside",
                                                "percent": 0
                                            },
                                            "qnet": {
                                                "barType": "qnet",
                                                "percent": 0
                                            },
                                            "language": {
                                                "barType": "language",
                                                "percent": 0
                                            },
                                            "intern": {
                                                "barType": "intern",
                                                "percent": 0
                                            },
                                            "sum": 0,
                                            "top": 100
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}